+++ src/db/storage/internetarchive.rs
@@ -40,6 +40,9 @@ impl Storage {
     /// Remove storages with given `ids`.
     /// Does not commit the transaction, you must do so yourself.
     pub async fn remove_by_file_ids(transaction: &mut Transaction<'_, Postgres>, file_ids: &[i64]) -> Result<()> {
+        if file_ids.is_empty() {
+            return Ok(());
+        }
         let stmt = "DELETE FROM stash.storage_internetarchive WHERE file_id = ANY($1::bigint[])";
         sqlx::query(stmt)
             .bind(file_ids)
@@ -50,6 +53,9 @@ impl Storage {
     /// Get internetarchive storage entities by exastash file ids.
     /// Entities which are not found will not be included in the resulting `Vec`.
     pub async fn find_by_file_ids(transaction: &mut Transaction<'_, Postgres>, file_ids: &[i64]) -> Result<Vec<Storage>> {
+        if file_ids.is_empty() {
+            return Ok(vec![]);
+        }
         // Note that we can get more than one row per unique file_id
         Ok(sqlx::query_as::<_, Storage>(
                 "SELECT file_id, ia_item, pathname, darked, last_probed
